Mod (./) win lose

+++ src/players/miss_hit.c
@@ -28,9 +28,10 @@ int miss_hit(pos_t *coords, char **map, char **map_enemy, int player)
     if (global.count == 8 &&
     (global.signal_value == HIT || global.signal_value == MISS)) {
         update_map(global.signal_value, *coords, map_enemy);
-        if (player == 2)
+        if (player == 2 && !global.victory)
             print_map(map, map_enemy);
-        send_signal(PLAY, global.pid);
+        if (!global.victory)
+            send_signal(PLAY, global.pid);
         my_putstr("waiting for enemy's attack...\n\n");
         global.count = 0;
         global.signal_value = 0;
